Refactor package structure and update imports for ptcolors
