+++ build_docs/src/conf.py
@@ -21,7 +21,7 @@ sys.path.insert(0, os.path.abspath("../../src/ptcolors"))
 
 # -- Project information -----------------------------------------------------
 
-project = "tcolors"
+project = "ptcolors"
 copyright = "2024, Elizabeth Consulting International Inc."
 author = "Elizabeth Consulting International Inc."
 
@@ -50,7 +50,7 @@ exclude_patterns: list[str] = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "alabaster"
+html_theme = "cloud"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
