+++ build_docs/src/ptcolors.rst
@@ -1,4 +1,7 @@
-.. automodule:: tcolors
+ptcolors
+========
+
+.. automodule:: ptcolors
    :members:
    :undoc-members:
    :show-inheritance:
